It now reads samples.
At least in the slow mode (the fast mode seems to get the wrong values). And the
scaling description is wired to the 1V/divsion setting until either jyetech updates
heir firmware, or I finish the settings mappings. And it doesn't exit properly.
And probably lots more. But hey - it output a sine wave!