Bug fixes.
Fix the silly error that was keeping it from reading values. This caused
the messages from the code to be seriously reworked and made a bit more
consistent in order to chase it down.

Fixed a bug that would have broken using multiple DSO112A's if it ever becomes
possible, as well as causing problems with buffering the IO better later on.
That caused a major code refactoring.